+++ src/core/window.c
@@ -1,16 +1,18 @@
 #include "core/window.h"
+#include "core/logger.h"
 
 #include <GLFW/glfw3.h>
 #include <stdio.h>
 
 void framebufferSizeCallback(GLFWwindow *window, int width, int height){
   glViewport(0, 0, width, height);
+  logMsg(DEBUG, "Changing window resolution to: %d, %d", width, height);
 }
 
 int initGLFW(void){
   if (!glfwInit()){
-    fprintf(stderr, "Failed to initialize GLFW.\n");
-    return -1; // for now, next version ill add a basic logger
+    logMsg(FAILURE, "Failed to initialize GLFW.");
+    return -1;
   }
  
   glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, GL_VER_MAJOR);
@@ -23,8 +25,8 @@ int initGLFW(void){
 int initWindow(GLFWwindow **window){
   *window = glfwCreateWindow(WIDTH, HEIGHT, TITLE, NULL, NULL);
 
-  if (*window == nullptr){
-    fprintf(stderr, "Failed to create GLFW window.\n");
+  if (*window == NULL){
+    logMsg(FAILURE, "Failed to create GLFW window.");
     return -1;
   }
 
@@ -32,8 +34,11 @@ int initWindow(GLFWwindow **window){
   glfwSwapInterval(1); // vsync enabled
 
   if (!gladLoadGLLoader((GLADloadproc) glfwGetProcAddress)){
-    fprintf(stderr, "Failed to intialize GLAD.\n");
+    logMsg(FAILURE, "Failed to intialize GLAD.");
     return -1;
+  }
+  else {
+    logMsg(SUCCESS, "GLAD initialized successfully");
   }
 
   glViewport(0, 0, WIDTH, HEIGHT);